+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### bug fixes
  * Do not change selects to the Bootstrap select class unless they are explicitly select fields. 
    Other types of controls also give select fields -- typeaheads and the like.
+ * Fix compatibility layer for Bootstrap selects when the `bootstrap-select-rails` is not installed.
 
 ## 1.2.0